Adds checkstyle plugin configuration and Wildfly rules with supress config
The setup should help the project to maintain certain standard of style closely associated to Wildfly.
The checkstyle-supressions.xml file is here to both streamline the configuration for anyone
and to add an exception to redundant final modifiers. I personally find it O.K. and it subjectively helps
me to navigate variables purposes in code. If you think the redundant final modifiers should be
removed, I remove them though.

Occurences are purely a matter of style IMHO:

[ERROR] Allocator.java:100:18: Redundant 'final' modifier. [RedundantModifier]
[ERROR] Allocator.java:192:21: Redundant 'final' modifier. [RedundantModifier]
[ERROR] Allocator.java:193:21: Redundant 'final' modifier. [RedundantModifier]
[ERROR] Allocator.java:194:21: Redundant 'final' modifier. [RedundantModifier]
[ERROR] DBAllocator.java:152:14: Redundant 'final' modifier. [RedundantModifier]

Supressed now.